server,webserver: now server handles encryption
This is a very big commit, notabily because server now handle the
extraction of data. It was necessary since the server own the encryption
keypair, so it have for responsibility to encrypt data.

This could be re-pushed in webserver in the future, in order to simplify
the server class (which is way too big currently)